Add files via upload
These are the pages that will appear to the user . The first image has a list of all the areas that he can rate, like parking. Each area has a rate button beside it,. Clicking on that button takes the user to image 2, where he has various factors that he can rate i.e lighting and so on. He has 4 options ( Y-Yes , N-No , NA-NotAnswered , N/A-Not Applicable) where yes is a postive response, no is a negative one, Not answered is when he cannot decide between Yes/No and not applicable is when he thinks that that particular factor doesn't influence his rating.